Move types.json into /utilities directory; update imports accordingly
